project fixes and authrepo

+++ PangWeb/Pages/Components/AdminPage/BlogComponents/AdminBlogComponent.razor.cs
@@ -16,9 +16,9 @@ namespace PangWeb.Pages.Components.AdminPage.BlogComponents
         }
 
         // Edit Blog dialog
-        private int SelectedId { get; set; } = 0;
+        private long SelectedId { get; set; } = 0;
 
-        private void SetSelectedId(int selectedRowId = 0)
+        private void SetSelectedId(long selectedRowId = 0)
         {
             SelectedId = selectedRowId;
         }
